add logentry

+++ WorkSpaceManager/views.py
@@ -31,23 +31,27 @@ def get_industrial_activity(request):
     })
 
 class WorkspaceManager(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated,IsWorkSpaceUser]
     jadoo_base_url = os.getenv("JADOO_BASE_URL")
 
     def delete(self,request,workspace_id):
         workspace_obj =get_object_or_404(WorkSpace,id=workspace_id)
-        if request.user.current_workspace_id == workspace_obj.id:
-            workspace_member= WorkspaceMember.objects.filter(user_account =request.user).first()
-            workspace_owner = WorkSpace.objects.filter(owner= request.user).first()
-            if workspace_owner:
-                request.user.current_workspace_id=workspace_owner.id
-            elif workspace_member:
+        if workspace_obj.owner == request.user:
+            if request.user.current_workspace_id == workspace_obj.id:
+                workspace_member= WorkspaceMember.objects.filter(user_account =request.user).first()
+                workspace_owner = WorkSpace.objects.filter(owner= request.user).first()
+                if workspace_owner:
+                    request.user.current_workspace_id=workspace_owner.id
+                elif workspace_member:
 
-                request.user.current_workspace_id = workspace_member.workspace.id
-            request.user.save()
+                    request.user.current_workspace_id = workspace_member.workspace.id
+                request.user.save()
 
-        workspace_obj.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            workspace_obj.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return  Response(status=status.HTTP_403_FORBIDDEN,data={
+            "message":"you dont have permission dont try !!!!!!"
+        })
 
 
     def get(self,request,workspace_id=None):
